Added health system
Added a health system, updated the HUD to have a heart counter. I am considering making the dashes also into a nice sprite but I lack the skill to do that so i'm just postponing it. I also added fall damage by popular demand (my girlfriend asked cuz her levels need it)
Also moved all of the killing logic into the player so that the killzone doesn't have to be called if the player dies or whatever reason.